+++ odoo/local-src/bso_restrict_attachments_visibility/__manifest__.py
@@ -1,0 +1,12 @@
+{
+    'name': 'BSO restrict attachments visibility',
+    'description': 'restrict access to attachments based on group users',
+    'author': 'BSO',
+    'website': 'https://www.bsonetwork.com',
+    'depends': ['base', 'hr'],
+    'data': [
+        'security/ir.model.access.csv',
+        'security/attachment_security.xml'
+    ],
+    'application': False,
+}